Remove deserializer_impl and move serialization_fixture to serializer_impl test.
